Backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
email-validator
fastapi-clerk-auth
pytest
python-jose
PyJWT
pytest-asyncio